Rules refactoring.
